Fix missing Toilet of Doom painting logic
Toilet of Doom can only be accessed from the Boss Arena Area, so also
requires the first paintings when paintings are shuffled.

To fix this, the Boss Arena Chest access rules have been moved into the
Boss Arena Area access rules, which is the parent of both Boss Arena
Chest and Toilet of Doom.